check point: at the point where if the trip time is less then the added times for driving and site tour times, start removing. Need to get the remove working

+++ Frank Lloyd Wright Trail/TripOrder.swift
@@ -12,12 +12,12 @@ class TripObject: NSObject {
     var startPoint: Double
     var endPoint: Double?
     var timeText: String?
-    var timeValue: Int?
+    var timeValue: Double?
     var distanceText: String?
     var distanceValue: Int?
     var image: UIImage?
     
-    init(startPoint: Double, endPoint: Double, timeText: String, timeValue: Int, distanceText: String, distanceValue: Int, image: UIImage?) {
+    init(startPoint: Double, endPoint: Double, timeText: String, timeValue: Double, distanceText: String, distanceValue: Int, image: UIImage?) {
         self.startPoint = startPoint
         self.endPoint = endPoint
         self.timeText = timeText
